refactor: reorganize domain structure and update app name
- Move user, role, and resource domains under new admin/ directory
- Update app name from CommitStreams to PlayWithLLM throughout codebase
- Remove unused pull request and repository domains
- Add BYPASS_AUTH_FOR_DEVELOPMENT config option
- Update MongoDB database name to pwllmdb
- Update deployment path to playwithllm-server
