+++ src/domains/admin/resource/api.js
@@ -1,7 +1,6 @@
 const express = require('express');
-const Joi = require('joi');
-const logger = require('../../libraries/log/logger');
-const { AppError } = require('../../libraries/error-handling/AppError');
+const logger = require('../../../libraries/log/logger');
+const { AppError } = require('../../../libraries/error-handling/AppError');
 
 const {
   create,
@@ -10,8 +9,6 @@ const {
   getById,
   updateById,
   deleteById,
-  getAllGroupedByType,
-  updateRolePermissions,
 } = require('./service');
 
 const {
@@ -20,11 +17,11 @@ const {
   idSchema,
   searchSchema,
 } = require('./request');
-const { validateRequest } = require('../../middlewares/request-validate');
-const { logRequest } = require('../../middlewares/log');
-const { isAuthorized } = require('../../middlewares/auth/authorization');
+const { validateRequest } = require('../../../middlewares/request-validate');
+const { logRequest } = require('../../../middlewares/log');
+const { isAuthorized } = require('../../../middlewares/auth/authorization');
 
-const model = 'Role';
+const model = 'Resource';
 
 const routes = () => {
   const router = express.Router();
@@ -36,6 +33,11 @@ const routes = () => {
     validateRequest({ schema: searchSchema, isQuery: true }),
     async (req, res, next) => {
       try {
+        console.log('req', {
+          url: req.url,
+          query: req.query,
+          originalUrl: req.originalUrl,
+        });
         const items = await search(req.query);
         res.json(items);
       } catch (error) {
