+++ src/domains/admin/resource/index.js
