+++ src/domains/admin/resource/service.js
@@ -1,6 +1,6 @@
-const logger = require('../../libraries/log/logger');
+const logger = require('../../../libraries/log/logger');
 const Model = require('./schema');
-const { AppError } = require('../../libraries/error-handling/AppError');
+const { AppError } = require('../../../libraries/error-handling/AppError');
 
 const model = 'resource';
