+++ src/domains/index.js
@@ -1,16 +1,8 @@
-const userRoutes = require('./user');
-const repositoryRoutes = require('./repository');
-const prRoutes = require('./pull');
-const roleRoutes = require('./role');
-const resourceRoutes = require('./resource');
+const adminRoutes = require('./admin');
 
-const defineRoutes = async (expressRouter) => {
-  
-  userRoutes(expressRouter);
-  repositoryRoutes(expressRouter);
-  prRoutes(expressRouter);
-  roleRoutes(expressRouter);
-  resourceRoutes(expressRouter);
+const defineRoutes = (expressRouter) => {
+  // Mount admin routes
+  adminRoutes(expressRouter);
 };
 
 module.exports = defineRoutes;
