--- src/domains/pull/index.js
@@ -1,7 +1,0 @@
-const { routes } = require('./api');
-
-const defineRoutes = (expressRouter) => {
-  expressRouter.use('/pulls', routes());
-};
-
-module.exports = defineRoutes;
