--- src/domains/repository/index.js
@@ -1,7 +1,0 @@
-const { routes } = require('./api');
-
-const defineRoutes = (expressRouter) => {
-  expressRouter.use('/repositories', routes());
-};
-
-module.exports = defineRoutes;
