+++ src/libraries/email/templates/verification.html
@@ -151,7 +151,7 @@
                     <!-- start copy -->
                     <tr>
                         <td align="left" bgcolor="#ffffff" style="padding: 24px; font-family: 'Source Sans Pro', Helvetica, Arial, sans-serif; font-size: 16px; line-height: 24px; border-bottom: 3px solid #d4dadf">
-                            <p style="margin: 0;">Best regards,<br>CommitStreams Team / Foyzul Karim</p>
+                            <p style="margin: 0;">Best regards,<br>PlayWithLLM Team / Foyzul Karim</p>
                         </td>
                     </tr>
                     <!-- end copy -->
